+++ OdiparPackPedidosModel/src/model/Paquete.java
@@ -2,9 +2,9 @@
 package com.example.backend.models;
 
 public class Paquete {
-    private int id_paquete;
-    private int fid_almacen;
-    private int fid_pedido;
-    private int fid_tramoActual;
+    private long id_paquete;
+    private long fid_almacen;
+    private long fid_pedido;
+    private long fid_tramoActual;
     private EstadoPaquete estado;
 }
