+++ OdiparPackPedidosModel/src/model/Pedido.java
@@ -4,9 +4,9 @@ package com.example.backend.models;
 import java.time.LocalDateTime;
 
 public class Pedido {
-    private int id_pedido;
-    private int fid_almacen;
-    private int fid_oficinaDest;
+    private long id_pedido;
+    private long fid_almacen;
+    private long fid_oficinaDest;
     private LocalDateTime fechaEntregaReal;
     private LocalDateTime fechaEntregaEstimada;
     private EstadoPedido estado;
