+++ OdiparPackPedidosModel/src/model/Simulacion.java
@@ -4,7 +4,7 @@ package com.example.backend.models;
 import java.time.LocalDateTime;
 
 public class Simulacion {
-    private int id_simulacion;
+    private long id_simulacion;
     private TipoSimulacion tipo;
     private LocalDateTime fechaInicioRealSimulacion;
     private LocalDateTime fechaInicioSimulacion;
