+++ OdiparPackRedLogisticaModel/src/model/Almacen.java
@@ -2,8 +2,8 @@
 package com.example.backend.models;
 
 public class Almacen {
-    private int id_almacen;
-    private int fid_ubicacion;
+    private long id_almacen;
+    private long fid_ubicacion;
     private int cantidadCamiones;
 
 }
