+++ OdiparPackRedLogisticaModel/src/model/Averia.java
@@ -4,7 +4,7 @@ package com.example.backend.models;
 import java.time.LocalDateTime;
 
 public class Averia {
-    private int id_averia;
+    private long id_averia;
     private TipoAveria tipo;
     private LocalDateTime fechaInicio;
     private LocalDateTime fechaFin;
