+++ OdiparPackRedLogisticaModel/src/model/Mantenimiento.java
@@ -5,7 +5,7 @@ import java.sql.Time;
 import java.time.LocalDateTime;
 
 public class Mantenimiento {
-    private int id_mantenimiento;
+    private long id_mantenimiento;
     private TipoMantenimiento tipo;
     private LocalDateTime fechaProgramada;
     private Time duracion;
