+++ OdiparPackRedLogisticaModel/src/model/Oficina.java
@@ -2,8 +2,8 @@
 package com.example.backend.models;
 
 public class Oficina {
-    private int id_oficina;
-    private int fid_ubicacion;
+    private long id_oficina;
+    private long fid_ubicacion;
     private int capacidadUtilizada;
     private int capacidadMaxima;
 }
