+++ OdiparPackRedLogisticaModel/src/model/TipoVehiculo.java
@@ -2,7 +2,7 @@
 package com.example.backend.models;
 
 public class TipoVehiculo {
-    private int id_tipoVehiculo;
+    private long id_tipoVehiculo;
     private String nombre;
     private float velocidad;
     private float kilometrajeMantenimiento;
