+++ OdiparPackRedLogisticaModel/src/model/VehiculoXRuta.java
@@ -2,8 +2,8 @@
 package com.example.backend.models;
 
 public class VehiculoXRuta {
-    private int id_vehiculoXRuta;
-    private int fid_camion;
-    private int fid_ruta;
+    private long id_vehiculoXRuta;
+    private long fid_camion;
+    private long fid_ruta;
     private float distanciaRecorrida;
 }
