+++ OdiparPackRutaModel/src/model/Bloqueo.java
@@ -4,7 +4,7 @@ package com.example.backend.models;
 import java.time.LocalDateTime;
 
 public class Bloqueo {
-    private int id_bloqueo;
+    private long id_bloqueo;
     private LocalDateTime fechaInicio;
     private LocalDateTime fechaFin;
     private Tramo tramoAfectado;
