+++ OdiparPackRutaModel/src/model/Region.java
@@ -2,7 +2,7 @@
 package com.example.backend.models;
 
 public class Region {
-    private int id_region;
+    private long id_region;
     private String nombre;
     private int diasLimite;
 }
