+++ OdiparPackRutaModel/src/model/Tramo.java
@@ -2,7 +2,7 @@
 package com.example.backend.models;
 
 public class Tramo {
-    private int id_tramo;
+    private long id_tramo;
     private Ubicacion origen;
     private Ubicacion destino;
     private boolean bloqueado;
