+++ OdiparPackRutaModel/src/model/Ubicacion.java
@@ -2,7 +2,7 @@
 package com.example.backend.models;
 
 public class Ubicacion {
-    private int id_ubicacion;
+    private long id_ubicacion;
     private String coordenada;
     private String ciudad;
     private Region region;
