+++ backend/src/main/java/com/example/backend/models/Paquete.java
@@ -1,5 +1,6 @@
 
 package com.example.backend.models;
+
 import jakarta.persistence.Id;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Column;
@@ -10,27 +11,27 @@ import jakarta.persistence.EnumType;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name="Paquete")
+@Table(name = "Paquete")
 public class Paquete {
-    
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_paquete")
-    private int id_paquete;
+    private long id_paquete;
 
     @Column(name = "fid_almacen")
-    private int fid_almacen;
+    private long fid_almacen;
 
     @Column(name = "fid_pedido")
-    private int fid_pedido;
+    private long fid_pedido;
 
     @Column(name = "fid_tramoActual")
-    private int fid_tramoActual;
+    private long fid_tramoActual;
 
     @Enumerated(EnumType.STRING)
     private EstadoPaquete estado;
 
-    public Paquete(int id_paquete, int fid_almacen, int fid_pedido, int fid_tramoActual, EstadoPaquete estado) {
+    public Paquete(long id_paquete, long fid_almacen, long fid_pedido, long fid_tramoActual, EstadoPaquete estado) {
         this.id_paquete = id_paquete;
         this.fid_almacen = fid_almacen;
         this.fid_pedido = fid_pedido;
