+++ backend/src/main/java/com/example/backend/models/Simulacion.java
@@ -4,9 +4,9 @@ package com.example.backend.models;
 import java.time.LocalDateTime;
 import jakarta.persistence.Table;
 
-@Table(name="Simulacion")
+@Table(name = "Simulacion")
 public class Simulacion {
-    private int id_simulacion;
+    private long id_simulacion;
     private TipoSimulacion tipo;
     private LocalDateTime fechaInicioRealSimulacion;
     private LocalDateTime fechaInicioSimulacion;
