+++ backend/src/main/java/com/example/backend/models/VehiculoXRuta.java
@@ -9,19 +9,19 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name="VehiculoXRuta")
+@Table(name = "VehiculoXRuta")
 public class VehiculoXRuta {
-    
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_vehiculoXRuta;
+    private long id_vehiculoXRuta;
 
-    @Column(name="fid_camion")
-    private int fid_camion;
+    @Column(name = "fid_camion")
+    private long fid_camion;
 
-    @Column(name="fid_ruta")
-    private int fid_ruta;
-    
-    @Column(name="distancia_recorrida")
+    @Column(name = "fid_ruta")
+    private long fid_ruta;
+
+    @Column(name = "distancia_recorrida")
     private float distanciaRecorrida;
 }
